lib/websocket/server/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
websockets==12.0
python-dotenv==1.0.0 